+++ tools/codegen/codespec/api_to_provider_spec_mapper.go
@@ -471,6 +471,7 @@ func apiSpecToDataSourcesModel(spec *high.Document, resourceConfig *config.Resou
 		}
 
 		ds.Singular = &Schema{
+			ExpandedModel:      dsConfig.SchemaOptions.ExpandedModel,
 			Description:        &oasReadOp.Description,
 			DeprecationMessage: resourceConfig.DeprecationMessage,
 			Discriminator:      singularDisc,
@@ -499,6 +500,7 @@ func apiSpecToDataSourcesModel(spec *high.Document, resourceConfig *config.Resou
 		}
 
 		ds.Plural = &Schema{
+			ExpandedModel:      dsConfig.SchemaOptions.ExpandedModel,
 			Description:        &oasListOp.Description,
 			DeprecationMessage: resourceConfig.DeprecationMessage,
 			Discriminator:      pluralDisc,